[SC-225247] move upload button (#1485)
* Move upload button

* move group dropdown

* linting
