+++ src/frontend/src/components/NavBar/components/RightNav/components/UserMenu/index.tsx
@@ -3,13 +3,9 @@ import { useState } from "react";
 import { API } from "src/common/api";
 import ENV from "src/common/constants/ENV";
 import { ROUTES } from "src/common/routes";
-import { StyledNavButton, StyledNavIconWrapper } from "./style";
+import { UserMenuButton } from "./style";
 
-interface UserMenuProps {
-  user: string | undefined;
-}
-
-const UserMenu = ({ user }: UserMenuProps): JSX.Element => {
+const UserMenu = (): JSX.Element => {
   const [anchorEl, setAnchorEl] = useState<Element | null>(null);
 
   const handleClick: React.MouseEventHandler<HTMLButtonElement> = (event) => {
@@ -30,17 +26,9 @@ const UserMenu = ({ user }: UserMenuProps): JSX.Element => {
 
   return (
     <>
-      <StyledNavButton
-        data-test-id="nav-user-menu"
-        onClick={handleClick}
-        endIcon={
-          <StyledNavIconWrapper>
-            <Icon sdsIcon="chevronDown" sdsSize="xs" sdsType="static" />
-          </StyledNavIconWrapper>
-        }
-      >
-        {user}
-      </StyledNavButton>
+      <UserMenuButton data-test-id="nav-user-menu" onClick={handleClick}>
+        <Icon sdsIcon="person" sdsSize="l" sdsType="static" />
+      </UserMenuButton>
       <Menu
         anchorEl={anchorEl}
         keepMounted
